refactoring js -> ts
sequelize의 js를 ts로 변경함
