+++ server/models/like.ts
@@ -1,20 +1,18 @@
 'use strict';
-const {
-  Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
+import { DataType, Model, ModelStatic, Sequelize } from 'sequelize';
+module.exports = (sequelize: Sequelize, DataTypes: { INTEGER: DataType; }) => {
   class Like extends Model {
     /**
      * Helper method for defining associations.
      * This method is not a part of Sequelize lifecycle.
      * The `models/index` file will call this method automatically.
      */
-    static associate(models) {
+    static associate(models: { [x:string]: ModelStatic<Model<any, any>>; }) {
       // define association here
       this.belongsTo(models.User, {
         foreignKey: 'user_id'
       });
-      this.belongsTo(models.Magainze, {
+      this.belongsTo(models.Magazine, {
         foreignKey: 'magazine_id'
       });
     }
