--- server/models/magainze_category.js
@@ -1,36 +1,0 @@
-'use strict';
-const {
-  Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class Magainze_Category extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      // define association here
-      this.belongsTo(models.Magainze, {
-        foreignKey: 'magazine_id'
-      });
-      this.belongsTo(models.Category, {
-        foreignKey: 'category_id'
-      });
-    }
-  };
-  Magainze_Category.init({
-    magazine_id: {
-      type: DataTypes.INTEGER,
-      allowNull: false
-    },
-    category_id: {
-      type: DataTypes.INTEGER,
-      allowNull: false
-    }
-  }, {
-    sequelize,
-    modelName: 'Magainze_Category',
-  });
-  return Magainze_Category;
-};
