+++ server/models/magazine_category.ts
@@ -1,44 +1,34 @@
 'use strict';
-const {
-  Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class Magazine_Feed extends Model {
+import { DataType, Model, ModelStatic, Sequelize } from 'sequelize';
+module.exports = (sequelize: Sequelize, DataTypes: { [x:string]: DataType; }) => {
+  class Magainze_Category extends Model {
     /**
      * Helper method for defining associations.
      * This method is not a part of Sequelize lifecycle.
      * The `models/index` file will call this method automatically.
      */
-    static associate(models) {
+    static associate(models: { [x:string]: ModelStatic<Model<any, any>>; }) {
       // define association here
       this.belongsTo(models.Magazine, {
         foreignKey: 'magazine_id'
       });
-      this.belongsTo(models.Feed, {
-        foreignKey: 'feed_id'
-      })
+      this.belongsTo(models.Category, {
+        foreignKey: 'category_id'
+      });
     }
   };
-  Magazine_Feed.init({
+  Magainze_Category.init({
     magazine_id: {
       type: DataTypes.INTEGER,
       allowNull: false
     },
-    feed_id: {
+    category_id: {
       type: DataTypes.INTEGER,
       allowNull: false
-    },
-    position_x: {
-      type: DataTypes.FLOAT,
-      allowNull: false
-    },
-    position_y: {
-      type: DataTypes.FLOAT,
-      allowNull: false
-    },
+    }
   }, {
     sequelize,
-    modelName: 'Magazine_Feed',
+    modelName: 'Magazine_Category',
   });
-  return Magazine_Feed;
+  return Magainze_Category;
 };
