+++ server/models/user.ts
@@ -1,15 +1,13 @@
 'use strict';
-const {
-  Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
+import { DataType, Model, Sequelize } from 'sequelize';
+module.exports = (sequelize: Sequelize, DataTypes: { [x:string]: DataType }) => {
   class User extends Model {
     /**
      * Helper method for defining associations.
      * This method is not a part of Sequelize lifecycle.
      * The `models/index` file will call this method automatically.
      */
-    static associate(models) {
+    static associate(models: any) {
       // define association here
     }
   };
