+++ server/package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "@types/node": "^14.11.8",
+    "@types/validator": "^13.1.0",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
@@ -22,6 +23,9 @@
     "typescript": "^4.0.3"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.8",
+    "@types/express": "^4.17.8",
+    "@types/express-session": "^1.17.0",
     "sequelize-cli": "^6.2.0"
   }
 }
