+++ server/tsconfig.json
@@ -1,6 +1,7 @@
 {
     "compilerOptions": {
-        "module": "CommonJS"
+        "module": "CommonJS",
+        "target": "es6",
     },
     "types": [
         "node"
